Handle list logdata
Summary:
The settings rollback incident logdata is a list instead of a dict. This was
causing an error to be thrown when pulling incidents that included a settings
rollback incident as we were expecting a dict.

To handle this, we know check for the type (list or dict) and handle each
case separately.

Test Plan:
Manually tested that pulling a list of incidents that included a
settings rollback does not break anymore (and that existing incidents still
behave as expected).